separation of graphql from db logic
