+++ src/db/orm/event.rs
@@ -1,9 +1,6 @@
-use super::Context;
 use crate::db::schema::*;
-use crate::db::orm::Project;
 use chrono::naive::NaiveDateTime;
 use diesel::{self, AsChangeset, Queryable};
-use juniper::FieldResult;
 use serde_derive::{Deserialize, Serialize};
 use slug::slugify;
 
@@ -50,39 +47,3 @@ impl EventInsert {
         }
     }
 }
-
-graphql_object!(Event: Context |&self| {
-    description: "A event, e.g. hackathon, code challenge, etc."
-
-    field id(&executor) -> i32 {
-        self.id
-    }
-
-    field name(&executor) -> &str {
-        &self.name
-    }
-
-    field start_time(&executor) -> &NaiveDateTime {
-        &self.start_time
-    }
-
-    field end_time(&executor) -> &NaiveDateTime {
-        &self.end_time
-    }
-
-    field description(&executor) -> Option<&str> {
-        match &self.description {
-            Some(desc) => Some(desc.as_str()),
-            None => None
-        }
-    }
-
-    field slug(&executor) -> &str {
-        &self.slug
-    }
-
-    field projects(&executor) -> FieldResult<Vec<Project>> {
-        let database: &diesel::PgConnection = &executor.context().database;
-        Ok(self.projects(database)?)
-    }
-});
