+++ src/db/orm/user.rs
@@ -1,12 +1,5 @@
-use super::Context;
-use crate::db::schema::*;
-use crate::db::schema::{projects, submissions};
-use crate::db::orm::{Project, Submission};
-use diesel::pg::expression::dsl::any;
-use diesel::BelongingToDsl;
-use diesel::ExpressionMethods;
-use diesel::QueryDsl;
-use diesel::RunQueryDsl;
+//use crate::db::schema::*;
+use crate::db::schema::{users};
 use diesel::{self, AsChangeset, Queryable};
 use serde_derive::{Deserialize, Serialize};
 
@@ -41,30 +34,6 @@ pub struct User {
     pub display_name: String,
     pub email: String,
 }
-
-graphql_object!(User: Context |&self| {
-    description: "A user"
-
-    field id(&executor) -> i32 {
-        self.id
-    }
-
-    field display_name(&executor) -> &str {
-        &self.display_name
-    }
-
-    field email(&executor) -> &str {
-        &self.email
-    }
-
-    field projects(&executor) -> Vec<Project> {
-        let database: &diesel::PgConnection = &executor.context().database;
-        let project_ids = Submission::belonging_to(self).select(submissions::project_id);
-        projects::table
-            .filter(projects::id.eq(any(project_ids)))
-            .load::<Project>(database).expect("Could not load projects")
-    }
-});
 
 #[derive(Serialize, Deserialize, GraphQLObject)]
 pub struct LoginRequest {
