+++ src/graphql/user.rs
@@ -1,0 +1,33 @@
+use crate::db::schema::{projects, submissions};
+use super::Context;
+use diesel::BelongingToDsl;
+use crate::types::{ Project, User, Submission };
+use diesel::pg::expression::dsl::any;
+use diesel::ExpressionMethods;
+use diesel::QueryDsl;
+use diesel::RunQueryDsl;
+
+graphql_object!(User: Context |&self| {
+    description: "A user"
+
+    field id(&executor) -> i32 {
+        self.id
+    }
+
+    field display_name(&executor) -> &str {
+        &self.display_name
+    }
+
+    field email(&executor) -> &str {
+        &self.email
+    }
+
+    field projects(&executor) -> Vec<Project> {
+        let database: &diesel::PgConnection = &executor.context().database;
+        let project_ids = Submission::belonging_to(self).select(submissions::project_id);
+        projects::table
+            .filter(projects::id.eq(any(project_ids)))
+            .load::<Project>(database).expect("Could not load projects")
+    }
+});
+
